+++ parsing/free.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   free.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yeeunpar <yeeunpar@student.42.fr>          +#+  +:+       +#+        */
+/*   By: woorikim <woorikim@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/14 14:02:17 by yeeunpar          #+#    #+#             */
-/*   Updated: 2024/03/19 00:18:14 by yeeunpar         ###   ########.fr       */
+/*   Updated: 2024/03/19 16:38:38 by woorikim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
