+++ parsing/map_validator.c
@@ -6,7 +6,7 @@
 /*   By: woorikim <woorikim@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/14 17:00:36 by yeeunpar          #+#    #+#             */
-/*   Updated: 2024/03/19 12:30:46 by woorikim         ###   ########.fr       */
+/*   Updated: 2024/03/19 16:42:18 by woorikim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static long	find_first_newline_index(char *map)
 {
-	long    i;
+	long	i;
 
 	i = 0;
 	while (map[i])
@@ -37,7 +37,7 @@ static char	*check_newlines_error(char	*map)
 	{
 		if (tmp[j] == '\n')
 			if (tmp[j + 1] == '\n' && tmp[j] != '\0' && tmp[j + 1])
-				print_error("Error : Map error\n");
+				print_error("Error\nMap error\n");
 		j++;
 	}
 	return (tmp);
@@ -52,12 +52,12 @@ void	map_validator(t_game *game)
 
 	height = 0;
 	if (game->player_cnt >= 2 || game->player_cnt == 0)
-		print_error("Error : player count error\n");
+		print_error("Error\nplayer count error\n");
 	location = find_first_newline_index(game->map);
 	copy_map = check_newlines_error(game->map + location);
 	map_2d = parsing_str(copy_map, '\n');
 	if (map_2d == (void *)0)
-		print_error("Error : Map malloc fail\n");
+		print_error("Error\nMap malloc fail\n");
 	checking_map2(map_2d);
 	checking_wall(map_2d);
 	game->map_2d = map_2d;
