+++ parsing/str.c
@@ -6,7 +6,7 @@
 /*   By: woorikim <woorikim@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/14 14:01:00 by yeeunpar          #+#    #+#             */
-/*   Updated: 2024/03/19 09:03:57 by woorikim         ###   ########.fr       */
+/*   Updated: 2024/03/19 16:44:12 by woorikim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,11 +36,11 @@ long	check_atoi(char *str)
 		if ('0' <= str[i] && str[i] <= '9')
 			res = (10 * res) + str[i] - '0';
 		else
-			print_error("Error : atoi invalid\n");
+			print_error("Error\natoi invalid\n");
 		i++;
 	}
 	if (i >= 11)
-		print_error("Error : Invaild range(long long)\n");
+		print_error("Error\nInvaild range(long long)\n");
 	return (res);
 }
 
@@ -60,14 +60,14 @@ int	str_ncompare(const char *str1, const char *str2, size_t n)
 
 int	set_atoi(char *str)
 {
-	int i;
-	int res;
+	int	i;
+	int	res;
 
 	i = 0;
-    res = check_atoi((str + i));
+	res = check_atoi((str + i));
 	if (str[i] == '-')
-		print_error("Error : Invaild_int\n");
+		print_error("Error\nInvaild_int\n");
 	if (!((0 <= res) && (res <= 255)))
-		print_error("Error : Invaild atoi range\n");
+		print_error("Error\nInvaild atoi range\n");
 	return (res);
 }
