+++ parsing/wall_vaildator.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   wall_vaildator.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yeeunpar <yeeunpar@student.42.fr>          +#+  +:+       +#+        */
+/*   By: woorikim <woorikim@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/18 23:20:46 by yeeunpar          #+#    #+#             */
-/*   Updated: 2024/03/19 00:18:56 by yeeunpar         ###   ########.fr       */
+/*   Updated: 2024/03/19 16:44:37 by woorikim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ static void	validate_first_last_line(char *map)
 	int	cnt;
 	int	first_wall_idx;
 
-    first_wall_idx = 0;
+	first_wall_idx = 0;
 	cnt = counting_str_length(map);
 	while (map[first_wall_idx])
 	{
@@ -40,7 +40,7 @@ static void	validate_first_last_line(char *map)
 		first_wall_idx++;
 	}
 	if (map[first_wall_idx] != '1' && map[cnt - 1] != '1')
-		print_error("Error : No Wall\n");
+		print_error("Error\nNo Wall\n");
 }
 
 int	check_all_spaces(char *map)
@@ -64,12 +64,12 @@ void	checking_wall(char **map)
 	int	i;
 	int	end_line;
 
-    i = -1;
+	i = -1;
 	end_line = 0;
 	while (map[end_line])
 		end_line++;
 	if (validate_end_line(map[0]) || validate_end_line(map[end_line - 1]))
-		print_error("Error : Invaild Wall\n");
+		print_error("Error\nInvaild Wall\n");
 	while (map[++i])
 	{
 		if (check_all_spaces(map[i]) == 1)
